requirements.txt
aiogram==3.8.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Для PostgreSQL (опционально)
pytest==8.0.0
python-dotenv==1.0.1